% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{Cheese data set}
\format{
A data frame with 16 observations on the following 4 variables.
}
\source{
Nateghi, L., Roohinejad, S., Totosaus, A., Mirhosseini, H., Shuhaimi, M., Meimandipour, A., Omidizadeh, A., Manap, M.Y.A. (2012). Optimization of textural properties and formulation of reduced fat Cheddar cheeses containing fat replacers. Journal of Food, Agriculture & Environment, 10(2), 46-54.
}
\usage{
data(cheese)
}
\description{
The Nateghi's cheese data set gives measurements of the cheddar cheese textures.
}
\details{
\itemize{
\item cohe The cohesiveness.
\item fat The fat percentage.
\item xangum The xanthan gum percentage.
\item sodcase The sodium caseinate percentage.
}
}
