% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{orange}
\alias{orange}
\title{Orange Data Set}
\format{
A data frame with 20 observations on the following 4 variables.
}
\source{
Mirhosseini, H., Tan, C.P. (2010). Discrimination of orange beverage emulsions with different formulations using multivariate analysis. Journal of the Science of Food and Agriculture, 90, 1308-1316.
}
\usage{
data(orange)
}
\description{
The Mirhosseini's and Tan's data set gives the measurements of the physicochemical properties of emulsion in orange drinks.
}
\details{
\itemize{
\item emulsion The emulsion density in g/cm3.
\item arabicgum The arabic gum quantity in g.
\item xanthangum The xanthan gum quantity in g.
\item orangeoil The orange oil quantity in g.
}
}
