% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SLRMss.R
\name{summary.SLRMss}
\alias{summary.SLRMss}
\title{SLRMss Object Summaries}
\usage{
\method{summary}{SLRMss}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SLRMss}.}

\item{...}{Currently ignored.}
}
\value{
A selected components extracted from a SLRMss object.
}
\description{
This function displays the summary for the fitted model. It includes parameter estimates under both, null and alternative hypothesis, corrected Akaike information criterion, Bayesian information criterion and the chosen statistics.
}
\examples{
data(orange)
fit <- SLRMss(emulsion ~ arabicgum + xanthangum + orangeoil, family="Student",
xi=3, testingbeta="xanthangum", statistic="LR", data=orange)
summary(fit)
}
